+++ tests/compile_run_golden.rs
@@ -42,6 +42,11 @@ const CASES: &[GoldenCase] = &[
         source_rel: "tests/sieve/sieve.err",
         expected: include_str!("compile_run_expected/sieve.expected"),
     },
+    GoldenCase {
+        name: "typeinfer",
+        source_rel: "tests/typeinfer/typeinfer.err",
+        expected: include_str!("compile_run_expected/typeinfer.expected"),
+    },
 ];
 
 fn parse_expected(content: &str) -> (i32, String) {
@@ -80,11 +85,65 @@ fn copy_fixture(manifest: &Path, rel: &Path, work: &Path, stem: &str) -> PathBuf
     dest
 }
 
+fn errand_binary() -> &'static str {
+    option_env!("CARGO_BIN_EXE_Errand").expect(
+        "CARGO_BIN_EXE_Errand must be set when running integration tests with `cargo test`",
+    )
+}
+
+/// Compile `case` with the `Errand` binary, run the resulting executable, and assert its
+/// stdout + exit code match the golden expectation. Panics (failing the test) on any mismatch.
+fn run_golden_case(case: &GoldenCase, errand: &str, manifest: &Path, work: &Path, trace: &Path) {
+    let (want_exit, want_stdout) = parse_expected(case.expected);
+
+    let src = copy_fixture(manifest, Path::new(case.source_rel), work, case.name);
+    let exe = work.join(format!("{}_exe", case.name));
+
+    let compile = Command::new(errand)
+        .current_dir(work)
+        .arg(&src)
+        .arg("-o")
+        .arg(&exe)
+        .arg("--trace-json")
+        .arg(trace)
+        .status()
+        .unwrap_or_else(|e| panic!("{}: spawn Errand: {e}", case.name));
+
+    assert!(
+        compile.success(),
+        "{}: compiler exited with {} (see {})",
+        case.name,
+        compile,
+        trace.display()
+    );
+
+    let run = Command::new(&exe)
+        .current_dir(work)
+        .output()
+        .unwrap_or_else(|e| panic!("{}: run {}: {e}", case.name, exe.display()));
+
+    let got_stdout = String::from_utf8_lossy(&run.stdout).into_owned();
+    let got_stderr = String::from_utf8_lossy(&run.stderr).into_owned();
+    let got_code = run.status.code();
+
+    assert_eq!(
+        got_code,
+        Some(want_exit),
+        "{}: exit code mismatch (stderr={got_stderr:?})",
+        case.name
+    );
+    assert_eq!(
+        got_stdout, want_stdout,
+        "{}: stdout mismatch\n--- expected ({} bytes) ---\n{want_stdout:?}\n--- got ({} bytes) ---\n{got_stdout:?}",
+        case.name,
+        want_stdout.len(),
+        got_stdout.len(),
+    );
+}
+
 #[test]
 fn compile_run_golden_fixtures() {
-    let errand = option_env!("CARGO_BIN_EXE_Errand").expect(
-        "CARGO_BIN_EXE_Errand must be set when running integration tests with `cargo test`",
-    );
+    let errand = errand_binary();
     let manifest = Path::new(env!("CARGO_MANIFEST_DIR"));
     let work = manifest.join("target").join("compile_run_golden");
     fs::create_dir_all(&work).expect("create work dir");
@@ -92,50 +151,28 @@ fn compile_run_golden_fixtures() {
     let trace = work.join("trace.jsonl");
 
     for case in CASES {
-        let (want_exit, want_stdout) = parse_expected(case.expected);
-
-        let src = copy_fixture(manifest, Path::new(case.source_rel), &work, case.name);
-        let exe = work.join(format!("{}_exe", case.name));
-
-        let compile = Command::new(errand)
-            .current_dir(&work)
-            .arg(&src)
-            .arg("-o")
-            .arg(&exe)
-            .arg("--trace-json")
-            .arg(&trace)
-            .status()
-            .unwrap_or_else(|e| panic!("{}: spawn Errand: {e}", case.name));
-
-        assert!(
-            compile.success(),
-            "{}: compiler exited with {} (see {})",
-            case.name,
-            compile,
-            trace.display()
-        );
-
-        let run = Command::new(&exe)
-            .current_dir(&work)
-            .output()
-            .unwrap_or_else(|e| panic!("{}: run {}: {e}", case.name, exe.display()));
-
-        let got_stdout = String::from_utf8_lossy(&run.stdout).into_owned();
-        let got_stderr = String::from_utf8_lossy(&run.stderr).into_owned();
-        let got_code = run.status.code();
-
-        assert_eq!(
-            got_code,
-            Some(want_exit),
-            "{}: exit code mismatch (stderr={got_stderr:?})",
-            case.name
-        );
-        assert_eq!(
-            got_stdout, want_stdout,
-            "{}: stdout mismatch\n--- expected ({} bytes) ---\n{want_stdout:?}\n--- got ({} bytes) ---\n{got_stdout:?}",
-            case.name,
-            want_stdout.len(),
-            got_stdout.len(),
-        );
+        run_golden_case(case, errand, manifest, &work, &trace);
     }
+}
+
+/// Development target for generics support: generic structs/enums and generic-returning
+/// functions do not yet compile + run end-to-end (SIR generation leaks unsolved
+/// existentials). The fixture documents the desired behavior; once it passes, remove
+/// `#[ignore]` and fold the case into `CASES`. Run on demand with:
+///   cargo test --test compile_run_golden -- --ignored
+#[test]
+#[ignore = "generics are not yet supported end-to-end; this is a development target"]
+fn compile_run_generics_target() {
+    let case = GoldenCase {
+        name: "generics_target",
+        source_rel: "tests/generics_target/generics_target.err",
+        expected: include_str!("compile_run_expected/generics_target.expected"),
+    };
+    let errand = errand_binary();
+    let manifest = Path::new(env!("CARGO_MANIFEST_DIR"));
+    let work = manifest.join("target").join("compile_run_golden");
+    fs::create_dir_all(&work).expect("create work dir");
+    let trace = work.join("trace.jsonl");
+
+    run_golden_case(&case, errand, manifest, &work, &trace);
 }
